Fundamental Kotlin - Generics.
